Version 0.4
New functionality : The app can now send SMS to the RATP in order to report aggression or sexual harassment. This functionality is also available for the city of Marseille but right now the app can only compose the number of the RTM because the RTM doesn't have an SMS platform yet.

Fixed a bug which can lead the app to always report "sexual" aggression in the city of paris

Import optimized, drawables too.