Version 0.4.3
Fixed #56 : a bug that caused the application to crash when the authentication server was not accessible or if the phone is in airplane mode (inside LoginActivity)

Optimization of application loading.
Fixed other minor bugs.

This update is mandatory because without it it is impossible to contact the application servers. (new anti-spam protection).

It will therefore be offered when the application is launched for each user.